# VSCODE AGENTS/Testing-Cursor_dev/Backend & Deploy/Foundation-Core/docs/contracts/openapi.foundation.yaml
openapi: 3.1.0
info:
  title: Foundation Core API
  version: 1.0.0
  description: Reusable backend-first runtime API for screenshot-driven frontend templates.
servers:
  - url: https://foundation-core.example.com
paths:
...
  /api/billing/checkout:
    post:
      summary: Create Stripe hosted checkout session using offer_key mapping
      responses:
        '200':
          description: Checkout URL envelope
        '400':
          description: Validation error
        '401':
          description: Billing auth required
        '404':
          description: Offer key not found
        '503':
          description: Billing not configured
...
  /api/media/upload:
    post:
      summary: Create a media upload intent
      responses:
        '200':
          description: Upload enabled with asset metadata and optional signed read URL
        '503':
          description: Storage adapter unavailable
...